requirements.txt
asgiref==3.7.2
Babel==2.14.0
Django==5.0
django-money==3.4.1
Faker==20.1.0
py-moneyed==3.0
python-dateutil==2.8.2
pytz==2023.3.post1
setuptools==69.0.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
validate-docbr==1.10.0